requirements.txt
PyYAML
langchain-core
python-dotenv
langchain-openai
streamlit
snowflake-connector-python